Interpolation own package and error handling [#165772818]
